+++ crates/client/src/client/types/components/get_all.rs
@@ -21,40 +21,4 @@ pub mod queries {
         use cynic::QueryBuilder;
         GetAllComponents::build(())
     }
-
-    pub mod api {
-        use std::sync::Arc;
-
-        use cynic::http::ReqwestExt;
-
-        use crate::client::types::components::get_all::queries::get_all_components_query;
-        use crate::client::ReactiveGraphClient;
-        use crate::client::ReactiveGraphClientExecutionError;
-        use crate::schema_graphql::types::component::Components as ComponentsVec;
-        use reactive_graph_graph::Component;
-
-        pub struct Components {
-            client: Arc<ReactiveGraphClient>,
-        }
-
-        impl Components {
-            pub fn new(client: Arc<ReactiveGraphClient>) -> Self {
-                Self { client }
-            }
-
-            pub async fn get_all_components(&self) -> Result<Option<Vec<Component>>, ReactiveGraphClientExecutionError> {
-                let components = self
-                    .client
-                    .client
-                    .post(self.client.url_graphql())
-                    .run_graphql(get_all_components_query())
-                    .await
-                    .map_err(ReactiveGraphClientExecutionError::FailedToSendRequest)?
-                    .data
-                    .map(|data| ComponentsVec(data.types.components))
-                    .map(From::from);
-                Ok(components)
-            }
-        }
-    }
 }