+++ crates/reactive-graph/src/client/handler.rs
@@ -2,6 +2,7 @@ use std::sync::Arc;
 
 use crate::client::commands::ClientCommands;
 use crate::client::instances::entities::entity_instances;
+use crate::client::instances::flows::flow_instances;
 use crate::client::instances::relations::relation_instances;
 use crate::client::result::CommandResult;
 use crate::client::system::command::execute_command;
@@ -11,6 +12,7 @@ use crate::client::system::remotes::remotes;
 use crate::client::system::shutdown::shutdown;
 use crate::client::types::components::components;
 use crate::client::types::entities::entity_types;
+use crate::client::types::flows::flow_types;
 use crate::client::types::relations::relation_types;
 use reactive_graph_client::ReactiveGraphClient;
 
@@ -22,13 +24,14 @@ pub(crate) async fn handle_command(client: &Arc<ReactiveGraphClient>, command: C
         ClientCommands::Plugins(args) => plugins(client, args).await,
         ClientCommands::Remotes(args) => remotes(client, args).await,
         ClientCommands::Shutdown => shutdown(client).await,
-        // Types
+        // Type System
         ClientCommands::Components(args) => components(client, args).await,
         ClientCommands::EntityTypes(args) => entity_types(client, args).await,
         ClientCommands::RelationTypes(args) => relation_types(client, args).await,
-        // TODO: Flow Types
+        ClientCommands::FlowTypes(args) => flow_types(client, args).await,
+        // Instance System
         ClientCommands::EntityInstances(args) => entity_instances(client, args).await,
         ClientCommands::RelationInstances(args) => relation_instances(client, args).await,
-        // TODO: Flow Instances
+        ClientCommands::FlowInstances(args) => flow_instances(client, args).await,
     }
 }