+++ crates/reactive-graph/src/client/types/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod components;
 pub(crate) mod entities;
 pub(crate) mod extension;
+pub(crate) mod flows;
 pub(crate) mod property_type;
 pub(crate) mod relations;