+++ crates/reactive-graph/src/client/types/property_type/output_format.rs
@@ -2,6 +2,6 @@ use crate::client::output_format::OutputFormatWrapper;
 
 pub(crate) type PropertyTypesOutputFormatWrapper = OutputFormatWrapper<
     reactive_graph_graph::PropertyType,
-    reactive_graph_table_model::types::property_type::PropertyType,
-    reactive_graph_table_model::types::property_type::PropertyTypesTableOptions,
+    reactive_graph_table_model::types::properties::PropertyType,
+    reactive_graph_table_model::types::properties::PropertyTypesTableOptions,
 >;