+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Interfaces: Provide JSON schemas of the type system and instance system via GraphQL API
+- Interfaces: Query for JSON Schemas of the type system and instance system via the Rust Client
+- CLI: Print the JSON Schemas of the type system and instance system
 
 ### Changed